Move annotation-box-container one level up.
