+++ src/sidebar-common/annotation-box/all-modes-footer.css
