+++ src/sidebar-common/annotation-box/all-modes-footer.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import Footer from '../footer'
+import { Footer } from '../components'
 import DefaultFooter from './default-footer'
 
 const styles = require('./all-modes-footer.css')
