+++ src/sidebar-common/annotation-box/annotation-box-container.css
