+++ src/sidebar-common/annotation-box/annotation-box-container.tsx
@@ -4,11 +4,11 @@ import moment from 'moment'
 import { connect } from 'react-redux'
 import noop from 'lodash/fp/noop'
 
-import * as actions from '../../sidebar/actions'
+import * as actions from '../sidebar/actions'
 import DefaultDeleteModeContent from './default-delete-mode-content'
 import EditModeContent from './edit-mode-content'
-import TruncatedTextRenderer from '../truncated-text-renderer'
-import { MapDispatchToProps } from '../../types'
+import { TruncatedTextRenderer } from '../components'
+import { MapDispatchToProps } from '../types'
 import { CrowdfundingBox } from 'src/common-ui/crowdfunding'
 import { remoteFunction } from 'src/util/webextensionRPC'
 import { EVENT_NAMES } from 'src/analytics/internal/constants'
