+++ src/sidebar-common/annotation-box/comment-tags.css
