+++ src/sidebar-common/annotation-box/default-delete-mode-content.tsx
