+++ src/sidebar-common/annotation-box/default-footer.css
