+++ src/sidebar-common/annotation-box/default-footer.tsx
