+++ src/sidebar-common/annotation-box/edit-mode-content.css
