+++ src/sidebar-common/annotation-box/index.tsx
